requirements.txt
# Web框架和相关依赖
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
Flask-Cors==4.0.0
gunicorn==21.2.0

# Excel处理
pandas==2.1.0
openpyxl==3.1.2
xlrd==2.0.1
numpy==1.26.0

# 数据库
SQLite3==2.6.0

# HTTP请求
requests==2.31.0
aiohttp==3.8.5
async-timeout==4.0.3

# 图片处理
Pillow==10.0.1

# 并发和异步处理
concurrent-futures==3.0.5
asyncio==3.4.3

# 日期和时间处理
python-dateutil==2.8.2

# JSON处理
simplejson==3.19.1

# 日志和错误处理
logging==0.4.9.6

# Base64和编码
pybase64==1.3.1

# API客户端库
openai==1.10.0  # OpenAI Python客户端库，用于流式输出支持

# 其他实用工具
tqdm==4.66.1